requirements.txt
shodan
colorama
urlopen
pyOpenSSL